+++ CHANGELOG.md
@@ -1,19 +1,21 @@
-# Version 0.1
+# Version 0.2
 
 ## 首版说明
 
 本版本包含以下内容：
 
-- 本模块支持tun、tproxy代理模式，使用TTLinkProxySelector.sh切换
+- 本模块支持tun、tproxy代理模式，使用TTLinkSelector.sh切换
 - 强化防火墙规则删除逻辑，防止使用过程中，防火墙规则重复引用
 - 隐藏OEM解锁选项
+- 代码合并到一个文件
   ro.oem_unlock_supported=0
 
 ## First Edition Notes
 
 This version includes the following:
 
-- This module supports tun and tproxy proxy modes, switched using TTLinkProxySelector.sh
+- This module supports tun and tproxy proxy modes, switched using TTLinkSelector.sh
 - Enhanced firewall rule deletion logic to prevent repeated references of firewall rules during use
 - Hide OEM unlock option
+- Merge code into a single file
   ro.oem_unlock_supported=0